add model chekpoint rank
