+++ pytorch_lightning/callbacks/model_checkpoint.py
@@ -91,6 +91,7 @@ class ModelCheckpoint(Callback):
                 f"Checkpoint directory {filepath} exists and is not empty with save_top_k != 0."
                 "All files in this directory will be deleted when a checkpoint is saved!"
             )
+        self._rank = 0
 
         self.monitor = monitor
         self.verbose = verbose
@@ -127,6 +128,16 @@ class ModelCheckpoint(Callback):
             mode = 'auto'
 
         self.monitor_op, self.kth_value, self.mode = mode_dict[mode]
+
+    @property
+    def rank(self) -> int:
+        """Process rank. In general, metrics should only be logged by the process with rank 0."""
+        return self._rank
+
+    @rank.setter
+    def rank(self, value: int) -> None:
+        """Set the process rank."""
+        self._rank = value
 
     def _del_model(self, filepath):
         try:
@@ -192,6 +203,7 @@ class ModelCheckpoint(Callback):
         filepath = os.path.join(self.dirpath, self.prefix + filename + str_ver + '.ckpt')
         return filepath
 
+    @rank_zero_only
     def on_validation_end(self, trainer, pl_module):
         # only run on main process
         if trainer.proc_rank != 0:
