+++ pytorch_lightning/loggers/tensorboard.py
@@ -13,8 +13,9 @@ import torch
 from pkg_resources import parse_version
 from torch.utils.tensorboard import SummaryWriter
 
-from pytorch_lightning.loggers.base import LightningLoggerBase, rank_zero_only
 from pytorch_lightning import _logger as log
+from pytorch_lightning.loggers.base import LightningLoggerBase
+from pytorch_lightning.utilities.distributed import rank_zero_only
 
 
 class TensorBoardLogger(LightningLoggerBase):
