+++ pytorch_lightning/loggers/trains.py
@@ -19,7 +19,8 @@ except ImportError:  # pragma: no-cover
                       ' install it with `pip install trains`.')
 
 from pytorch_lightning import _logger as log
-from pytorch_lightning.loggers.base import LightningLoggerBase, rank_zero_only
+from pytorch_lightning.loggers.base import LightningLoggerBase
+from pytorch_lightning.utilities.distributed import rank_zero_only
 
 
 class TrainsLogger(LightningLoggerBase):
