+++ pytorch_lightning/utilities/distributed.py
@@ -1,0 +1,17 @@
+from functools import wraps
+from typing import Callable
+
+
+def rank_zero_only(fn: Callable):
+    """Decorate a logger method to run it only on the process with rank 0.
+
+    Args:
+        fn: Function to decorate
+    """
+
+    @wraps(fn)
+    def wrapped_fn(self, *args, **kwargs):
+        if self.rank == 0:
+            fn(self, *args, **kwargs)
+
+    return wrapped_fn
