+++ tests/loggers/test_base.py
@@ -5,7 +5,8 @@ import numpy as np
 
 import tests.base.utils as tutils
 from pytorch_lightning import Trainer
-from pytorch_lightning.loggers import LightningLoggerBase, rank_zero_only, LoggerCollection
+from pytorch_lightning.loggers import LightningLoggerBase, LoggerCollection
+from pytorch_lightning.utilities.distributed import rank_zero_only
 from tests.base import LightningTestModel
